+++ lib/pmdtester/builders/html_report_builder.rb
@@ -19,6 +19,7 @@ module PmdTester
       index.puts html_report
       index.close
 
+      puts "Built difference report of #{project.name} successfully!"
       report_dir
     end